feat: send `'planned'`, `'result'`, and `'done'` messages (#4)
Send `'planned'` and `'done'` messages along with the `'result'` messages.

The `'planned'` message type summarizes the number of tests that will run. This message might come in handy in the future for reporters (e.g. a tap reporter).

The `'done'` message type is sent when a suite’s tests are done. This message is useful as there might be more than one suite in a file, so a reporter that aggregates its output can send the suite’s logs when it ends, instead of waiting for all of them to finish.

PR #4